PP-3678 Move payments resource validation tests into own package
with @simad
