+++ src/test/java/uk/gov/pay/api/it/validation/PaymentsResourceDescriptionValidationITest.java
@@ -1,10 +1,11 @@
-package uk.gov.pay.api.it;
+package uk.gov.pay.api.it.validation;
 
 import com.jayway.jsonassert.JsonAssert;
 import com.jayway.restassured.response.ValidatableResponse;
 import org.apache.commons.lang3.RandomStringUtils;
 import org.junit.Before;
 import org.junit.Test;
+import uk.gov.pay.api.it.PaymentResourceITestBase;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -15,7 +16,7 @@ import static javax.ws.rs.core.HttpHeaders.AUTHORIZATION;
 import static org.hamcrest.Matchers.hasSize;
 import static org.hamcrest.core.Is.is;
 
-public class PaymentsResourceAgreementIdValidationITest extends PaymentResourceITestBase {
+public class PaymentsResourceDescriptionValidationITest extends PaymentResourceITestBase {
 
     @Before
     public void setUpBearerToken() {
@@ -23,13 +24,13 @@ public class PaymentsResourceAgreementIdValidationITest extends PaymentResourceI
     }
 
     @Test
-    public void createPayment_responseWith400_whenAgreementIdIsNumeric() throws IOException {
+    public void createPayment_responseWith400_whenDescriptionIsNumeric() throws IOException {
 
         String payload = "{" +
                 "  \"amount\" : 9900," +
+                "  \"description\" : 1234," +
                 "  \"reference\" : \"Some reference\"," +
-                "  \"description\" : \"Some description\"," +
-                "  \"agreement_id\" : 1234" +
+                "  \"return_url\" : \"https://example.com\"" +
                 "}";
 
         InputStream body = postPaymentResponse(API_KEY, payload)
@@ -40,19 +41,19 @@ public class PaymentsResourceAgreementIdValidationITest extends PaymentResourceI
 
         JsonAssert.with(body)
                 .assertThat("$.*", hasSize(3))
-                .assertThat("$.field", is("agreement_id"))
+                .assertThat("$.field", is("description"))
                 .assertThat("$.code", is("P0102"))
-                .assertThat("$.description", is("Invalid attribute value: agreement_id. Must be a valid agreement ID"));
+                .assertThat("$.description", is("Invalid attribute value: description. Must be a valid string format"));
     }
 
     @Test
-    public void createPayment_responseWith400_whenAgreementIdIsEmpty() throws IOException {
+    public void createPayment_responseWith400_whenDescriptionIsEmpty() throws IOException {
 
         String payload = "{" +
                 "  \"amount\" : 9900," +
+                "  \"description\" : \"\"," +
                 "  \"reference\" : \"Some reference\"," +
-                "  \"description\" : \"Some description\"," +
-                "  \"agreement_id\" : \"\"" +
+                "  \"return_url\" : \"https://example.com\"" +
                 "}";
 
         InputStream body = postPaymentResponse(API_KEY, payload)
@@ -63,19 +64,19 @@ public class PaymentsResourceAgreementIdValidationITest extends PaymentResourceI
 
         JsonAssert.with(body)
                 .assertThat("$.*", hasSize(3))
-                .assertThat("$.field", is("agreement_id"))
+                .assertThat("$.field", is("description"))
                 .assertThat("$.code", is("P0101"))
-                .assertThat("$.description", is("Missing mandatory attribute: agreement_id"));
+                .assertThat("$.description", is("Missing mandatory attribute: description"));
     }
 
     @Test
-    public void createPayment_responseWith400_whenAgreementIdIsBlank() throws IOException {
+    public void createPayment_responseWith400_whenDescriptionIsBlank() throws IOException {
 
         String payload = "{" +
                 "  \"amount\" : 9900," +
+                "  \"description\" : \"    \"," +
                 "  \"reference\" : \"Some reference\"," +
-                "  \"description\" : \"Some description\"," +
-                "  \"agreement_id\" : \"    \"" +
+                "  \"return_url\" : \"https://example.com\"" +
                 "}";
 
         InputStream body = postPaymentResponse(API_KEY, payload)
@@ -86,19 +87,18 @@ public class PaymentsResourceAgreementIdValidationITest extends PaymentResourceI
 
         JsonAssert.with(body)
                 .assertThat("$.*", hasSize(3))
-                .assertThat("$.field", is("agreement_id"))
+                .assertThat("$.field", is("description"))
                 .assertThat("$.code", is("P0101"))
-                .assertThat("$.description", is("Missing mandatory attribute: agreement_id"));
+                .assertThat("$.description", is("Missing mandatory attribute: description"));
     }
-    
+
     @Test
-    public void createPayment_responseWith400_whenAgreementIdIsNull() throws IOException {
+    public void createPayment_responseWith400_whenDescriptionIsMissing() throws IOException {
 
         String payload = "{" +
                 "  \"amount\" : 9900," +
                 "  \"reference\" : \"Some reference\"," +
-                "  \"description\" : \"Some description\"," +
-                "  \"agreement_id\" : null" +
+                "  \"return_url\" : \"https://example.com\"" +
                 "}";
 
         InputStream body = postPaymentResponse(API_KEY, payload)
@@ -109,21 +109,45 @@ public class PaymentsResourceAgreementIdValidationITest extends PaymentResourceI
 
         JsonAssert.with(body)
                 .assertThat("$.*", hasSize(3))
-                .assertThat("$.field", is("agreement_id"))
+                .assertThat("$.field", is("description"))
                 .assertThat("$.code", is("P0101"))
-                .assertThat("$.description", is("Missing mandatory attribute: agreement_id"));
+                .assertThat("$.description", is("Missing mandatory attribute: description"));
     }
 
     @Test
-    public void createPayment_responseWith422_whenAgreementIdSizeIsGreaterThanMaxLength() throws IOException {
+    public void createPayment_responseWith400_whenDescriptionIsNull() throws IOException {
 
-        String aTooLongAgreementId = RandomStringUtils.randomAlphanumeric(27);
 
         String payload = "{" +
                 "  \"amount\" : 9900," +
+                "  \"description\" : null," +
                 "  \"reference\" : \"Some reference\"," +
-                "  \"description\" : \"Some description\"," +
-                "  \"agreement_id\" : \"" + aTooLongAgreementId + "\"" +
+                "  \"return_url\" : \"https://example.com\"" +
+                "}";
+
+        InputStream body = postPaymentResponse(API_KEY, payload)
+                .statusCode(400)
+                .contentType(JSON)
+                .extract()
+                .body().asInputStream();
+
+        JsonAssert.with(body)
+                .assertThat("$.*", hasSize(3))
+                .assertThat("$.field", is("description"))
+                .assertThat("$.code", is("P0101"))
+                .assertThat("$.description", is("Missing mandatory attribute: description"));
+    }
+
+    @Test
+    public void createPayment_responseWith422_whenDescriptionSizeIsGreaterThanMaxLength() throws IOException {
+
+        String aVeryLongReference = RandomStringUtils.randomAlphanumeric(256);
+
+        String payload = "{" +
+                "  \"amount\" : 9900," +
+                "  \"description\" : \"" + aVeryLongReference + "\"," +
+                "  \"reference\" : \"Some reference\"," +
+                "  \"return_url\" : \"https://www.example.com/return_url\"" +
                 "}";
 
         InputStream body = postPaymentResponse(API_KEY, payload)
@@ -134,19 +158,19 @@ public class PaymentsResourceAgreementIdValidationITest extends PaymentResourceI
 
         JsonAssert.with(body)
                 .assertThat("$.*", hasSize(3))
-                .assertThat("$.field", is("agreement_id"))
+                .assertThat("$.field", is("description"))
                 .assertThat("$.code", is("P0102"))
-                .assertThat("$.description", is("Invalid attribute value: agreement_id. Must be less than or equal to 26 characters length"));
+                .assertThat("$.description", is("Invalid attribute value: description. Must be less than or equal to 255 characters length"));
     }
 
     @Test
-    public void createPayment_responseWith400_whenAgreementIdHasNotAValidJsonValue() throws IOException {
+    public void createPayment_responseWith400_whenDescriptionHasNotAValidJsonValue() throws IOException {
 
         String payload = "{" +
                 "  \"amount\" : 9900," +
+                "  \"description\" : " +
                 "  \"reference\" : \"Some reference\"," +
-                "  \"description\" : \"Some description\"," +
-                "  \"agreement_id\" : " +
+                "  \"return_url\" : \"https://example.com\"" +
                 "}";
 
         InputStream body = postPaymentResponse(API_KEY, payload)
@@ -162,13 +186,13 @@ public class PaymentsResourceAgreementIdValidationITest extends PaymentResourceI
     }
 
     @Test
-    public void createPayment_responseWith400_whenAgreementIdFieldIsNotExpectedJsonField() throws IOException {
+    public void createPayment_responseWith400_whenDescriptionFieldIsNotExpectedJsonField() throws IOException {
 
         String payload = "{" +
                 "  \"amount\" : 9900," +
+                "  \"description\" : {\"whatever\" : 1}," +
                 "  \"reference\" : \"Some reference\"," +
-                "  \"description\" : \"Some description\"," +
-                "  \"agreement_id\" : []" +
+                "  \"return_url\" : \"https://example.com\"" +
                 "}";
 
         InputStream body = postPaymentResponse(API_KEY, payload)
@@ -179,9 +203,9 @@ public class PaymentsResourceAgreementIdValidationITest extends PaymentResourceI
 
         JsonAssert.with(body)
                 .assertThat("$.*", hasSize(3))
-                .assertThat("$.field", is("agreement_id"))
+                .assertThat("$.field", is("description"))
                 .assertThat("$.code", is("P0102"))
-                .assertThat("$.description", is("Invalid attribute value: agreement_id. Must be a valid agreement ID"));
+                .assertThat("$.description", is("Invalid attribute value: description. Must be a valid string format"));
     }
 
     private ValidatableResponse postPaymentResponse(String bearerToken, String payload) {
